Full code
yeah. all of it. sorry. i made it all in quite a frenzy. implements state design pattern for pages, template design pattern, and aspects of observer. also naturally uses mvc. threading also, file manipulation, used text files to act as both a stand in for the database as well as for certain input methods such as the kiosk scanner. the visuals are pretty rough but i'm going to go ahead and blame that on the api. fairly uncommented but i tried to keep decent naming conventions.